[IMP] sale_amazon: add information about compliance with Amazon's security standards
Depending on whether the seller made a request to get access to Personally Identifiable Information of its customers, and on other factors, Amazon may require him to fill out a second form about his data protection policy.
This commit makes it clear that compliance with Amazon's security standards is not guaranteed if the seller applying for Amazon MWS access keys is hosted on Odoo (SaaS or .sh).

Additionally, it add a section in which the questions asked to Amazon sellers are listed and answered for the Odoo Online and Odoo.sh platforms.